requirements.txt
Django==4.2
python-dotenv==1.0.1
Markdown==3.6
requests==2.31.0
slackclient==2.9.4
langchain==0.2.1
langchain-google-genai==1.0.5
pytz==2024.1